+++ sources/Device/src/Ampermeter/Calculator/ResolverAC.h
@@ -8,7 +8,7 @@ class ResolverAC
 {
 public:
     ResolverAC(const BufferADC &, const Period &);
-    double GetResult() const { return result; }
+    float GetResult() const { return result; }
 private:
-    double result = 0.0;
+    float result = 0.0f;
 };
